+++ CHANGES.rst
@@ -16,6 +16,10 @@ Changelog of threedigrid-builder
 
 - Added cross section width, height, shape to lines.
 
+- Added sewerage_type (pipes) and sewerage (weirs/orifices) to lines.
+
+- Fix: accept unknown sewerage types.
+
 
 0.16.0 (2022-01-06)
 -------------------